Blender_3DC_UE_Tools.py
Blender Script in Addition to the 3DCoat Applink. It ungroups the 3DC_Group_node and links the PBR Textures so that it can be used with the Unreal Pipeline. 
Albedo, Roughness, Metalness and Normals get exported with Textures in UE.  This script adds some Buttons in the Shader Editor. The first Button removes unwanted Material Duplicates. The second ungroups the Group Node and the last relinks to the channels Unreal understands.